Updated printGrade function and classes variable
This commit has moved the printing of the grade out to a separate function. Additionally, the value of "classes" was previously being updated before the key was being checked, so in instances where people would enter a non-grade value, this would increment "classes" by one which resulted in GPA being incorrect. This has been changed to only add a class in instances where a valid grade was entered.