Update GPACalculator.java
Updated with an MVP solution so that calculating a GPA will work when "Submit" button is pushed. GPA will update each time this happens. 

Next, I want to get the parseInput() method to update gradeTotal and subjectCount directly. This means the two ActionListened events (Reset and Submit) will be able to call these directly and it'll be able to update the GPA count that's displayed to customers.

Reset button doesn't currently work.