undoAddition() and error correction
Implemented undoAddition(), which will undo the addition of the last added grade in case the student made a mistake adding a particular grade. Also updated .getText() in addGradeButton's actionListener so that it will make the input into an upper case character, in case the student types "a" instead of "A".